General refactor and new auto test build with cmake
