+++ .gitignore
@@ -1,9 +1,6 @@
 .vscode
-Build/
-bin/
-lib/
-release/
 CMakeFiles/
+Out/
 
 CMakeCache.txt
 Makefile
