+++ Include/Utils.hpp
@@ -1,0 +1,15 @@
+#ifndef UTILS_HPP
+#define UTILS_HPP
+
+#include <tuple>
+#include <vector>
+
+
+namespace Utils
+{
+    int GetTerminalWidth(void);
+    int GetTerminalHeight(void);
+    int GetProgressiveId(void);
+}
+
+#endif
