--- src/utils.hpp
@@ -1,16 +1,0 @@
-#ifndef UTILS_HPP
-#define UTILS_HPP
-
-#include <tuple>
-#include <vector>
-#include "window.hpp"
-
-
-namespace UTILS
-{
-    int getTerminalWidth(void);
-    int getTerminalHeight(void);
-    int getProgressiveId(void);
-}
-
-#endif
